FLytt stubkode til egen fil
